Started Objectives
created and objectives object that when in contact with the local player
will tell the server that someone is stood on it.
once the server processes this and sends it back it is handled by the
new incoming packet script for case 15
this updates the objectiveA, B or C Object with the calculated score
devised by the server based on the updates its received.
It then draws to the GUI a circular health bar ( new script ) that will
display the capture progress ot the player.